re-locate custodian icon
